requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
pytorch-lightning
pinecone
pydantic
pydantic-settings
requests
python-multipart
pytest
httpx
uuid
